Added user intergration

+++ client/src/components/Admin/Main/Quizzes/CreateQuiz/AddQuiz/AddQuiz.jsx
@@ -6,14 +6,19 @@ import InputReducer from "../../../../../Reducer/InputReducer";
 import SelectReducer from "../../../../../Reducer/SelectReducer/SelectReducer";
 import "./AddQuiz.css";
 const AddQuiz = () => {
-  const [question, setQuestion] = useState("");
-  const [options, setOptions] = useState({
-    A: "",
-    B: "",
-    C: "",
-    D: "",
-    correctOption: "",
+  const [question, setQuestion] = useState({
+    statement: "",
+    options: {
+      A: "",
+      B: "",
+      C: "",
+      D: "",
+      correctOption: "",
+    },
+    type_id: "mcq",
   });
+
+  const [updateFlag, setUpdateFlag] = useState(false);
   const [authState] = useContext(AuthContext);
   const [questions, setQuestions] = useState([]);
   const { id } = useParams();
@@ -40,7 +45,7 @@ const AddQuiz = () => {
         data: { mcq },
         status,
       } = await helperService.createQuizQuestion(
-        { quiz_id: id, statement: question, options, type_id: "mcq" },
+        { quiz_id: id, ...question },
         { headers: { Authorization: authState.user.token } }
       );
       if (status === 201) {
@@ -50,16 +55,64 @@ const AddQuiz = () => {
       console.log(err);
     }
   };
-  const deleteQuestion = async () => {
+  const deleteQuestion = async (question) => {
     try {
-      const { data, status } = await helperService.deleteQuestion();
+      const { data, status } = await helperService.deleteQuestion({...question,type_id:"mcq"},{headers:{Authorization:authState.user.token}});
+      if(status === 202){
+        console.log(questions,question)
+        setQuestions(questions.filter((ques)=>ques.question_id !== question.question_id))
+      }
     } catch (err) {
       console.log(err);
     }
   };
+  const updateQuestion = async () => {
+    try {
+      const { status } = await helperService.updateQuestion(
+        { ...question, quiz_id: id },
+        { headers: { Authorization: authState.user.token } }
+      );
+      if (status === 200) {
+        console.log(
+          questions.map((ques) => {
+            if (ques.question_id === question.question_id) return question;
+            return ques;
+          })
+        );
+        setQuestions(
+          questions.map((ques) => {
+            if (ques.question_id === question.question_id) return question;
+            return ques;
+          })
+        );
+      }
+    } catch (err) {
+      console.log(err);
+    } finally {
+      setUpdateFlag(false);
+      setQuestion({
+        statement: "",
+        options: {
+          A: "",
+          B: "",
+          C: "",
+          D: "",
+          correctOption: "",
+        },
+        type_id: "mcq",
+      });
+    }
+  };
+
+  const updateDetails = async (question) => {
+    setUpdateFlag(true);
+    setQuestion({ ...question, type_id: "mcq" });
+  };
+
   useEffect(() => {
     fetchQuizzes();
   }, []);
+
   return (
     <div className="container" style={{ height: "100vh", overflowY: "scroll" }}>
       <p className="text-left dash-title-category pb-2 mt-5">Add Quiz</p>
@@ -74,8 +127,10 @@ const AddQuiz = () => {
         multiline
         rows={10}
         variant="outlined"
-        value={question}
-        onClickHandler={(value) => setQuestion(value)}
+        value={question.statement}
+        onClickHandler={(value) =>
+          setQuestion({ ...question, statement: value })
+        }
       />
       <div className="d-flex">
         <div className="col-md-6">
@@ -87,8 +142,13 @@ const AddQuiz = () => {
             multiline
             rows={2}
             variant="outlined"
-            value={options.A}
-            onClickHandler={(value) => setOptions({ ...options, A: value })}
+            value={question.options.A}
+            onClickHandler={(value) =>
+              setQuestion({
+                ...question,
+                options: { ...question.options, A: value },
+              })
+            }
           />
         </div>
         <div className="col-md-6">
@@ -100,8 +160,13 @@ const AddQuiz = () => {
             multiline
             rows={2}
             variant="outlined"
-            value={options.B}
-            onClickHandler={(value) => setOptions({ ...options, B: value })}
+            value={question.options.B}
+            onClickHandler={(value) =>
+              setQuestion({
+                ...question,
+                options: { ...question.options, B: value },
+              })
+            }
           />
           {/* <TextField
            
@@ -117,8 +182,13 @@ const AddQuiz = () => {
             fullWidth
             multiline
             rows={2}
-            value={options.C}
-            onClickHandler={(value) => setOptions({ ...options, C: value })}
+            value={question.options.C}
+            onClickHandler={(value) =>
+              setQuestion({
+                ...question,
+                options: { ...question.options, C: value },
+              })
+            }
           />
         </div>
         <div className="col-md-6">
@@ -130,8 +200,13 @@ const AddQuiz = () => {
             multiline
             rows={2}
             variant="outlined"
-            value={options.D}
-            onClickHandler={(value) => setOptions({ ...options, D: value })}
+            value={question.options.D}
+            onClickHandler={(value) =>
+              setQuestion({
+                ...question,
+                options: { ...question.options, D: value },
+              })
+            }
           />
         </div>
         {/* <div className="col-md-6"> */}
@@ -144,15 +219,29 @@ const AddQuiz = () => {
             array={["A", "B", "C", "D"]}
             className="w-100 mt-3"
             name="Correct Option"
-            value={options.correctOption}
+            value={question.options.correctOption}
             handleSelect={(e) =>
-              setOptions({ ...options, correctOption: e.target.value })
+              setQuestion({
+                ...question,
+                options: { ...question.options, correctOption: e.target.value },
+              })
             }
           />
         </div>
         <div className="create-con mt-4 clearfix">
-          <button className="p-2 float-right" onClick={createQuestion}>
-            ADD QUIZ<i className="fas fa-plus pr-2 pl-3"></i>
+          <button
+            className="p-2 float-right"
+            onClick={updateFlag ? updateQuestion : createQuestion}
+          >
+            {updateFlag ? (
+              <span>
+                Update Quiz<i className="fas fa-plus pr-2 pl-3"></i>
+              </span>
+            ) : (
+              <span>
+                ADD Quiz<i className="fas fa-plus pr-2 pl-3"></i>
+              </span>
+            )}
           </button>
         </div>
       </div>
@@ -161,8 +250,11 @@ const AddQuiz = () => {
           <div className="respective-question-badge col-md-4 mt-2 mb-2 p-3">
             <div className="d-flex flex-column">
               <div className="edit-add-quiz p-2 d-flex align-items-end justify-content-end">
-                <i className="fas fa-edit mr-2 ml-2"></i>
-                <i className="fas fa-trash mr-2 ml-2"></i>
+                <i
+                  className="fas fa-edit mr-2 ml-2"
+                  onClick={() => updateDetails(question)}
+                ></i>
+                <i className="fas fa-trash mr-2 ml-2" onClick={() => deleteQuestion(question)}></i>
               </div>
               <div className="d-flex">
                 <span className="add-quiz-question-span mt-2 mb-2">
